requirements.txt
py-algorand-sdk==1.20.2
pyteal==0.19.0
python-dotenv==1.0.0